Update README.md
The prior text only had a heading and the description of the file's contents. I changed the heading and description while adding an emoji to the description, made use of a fenced code block which houses brief information about me, provided task list and made use of a strikethrough on the closing sentence.